Refactor validation into separate decorator class
